bird_detection_station/requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
attrs==19.3.0
black==21.6b0
certifi==2020.4.5.2
cffi==1.12.3
chardet==3.0.4
click==7.1.2
fastapi==0.58.0
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.1
mock==4.0.3
more-itertools==8.4.0
mypy==0.910
mypy-extensions==0.4.3
numpy==1.17.2
packaging==20.4
paho-mqtt==1.5.0
pluggy==0.13.1
py==1.8.2
pycparser==2.19
pydantic==1.5.1
pynmea2==1.15.0
pyparsing==2.4.7
pyserial==3.4
pytest==5.4.3
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
sounddevice==0.3.13
starlette==0.13.4
types-PyYAML==5.4.3
types-requests==2.25.0
typing-extensions==3.10.0.0
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.4
websockets==8.1
zipp==3.1.0
